+++ game.py
@@ -1,56 +1,67 @@
 from player import Player
 import ship
-# import sys
+from collections import OrderedDict
 
 
 def play():
     """ Input for character movement and accessing inventory"""
     print("Escape from the Ship!")
-    # print("""
-    #       Do you want to go forward, aftward, port, starboard, or check
-    #       inventory
-    #       """)
     player = Player()
     # Possible player directions and actions
     while True:
         # define the player's start position
         position = ship.tile_at(player.x, player.y)
+        print(player.x, player.y)
         # print the intro at the start position
         print(position.intro_text())
         # modify health points of player when attacked
         position.modify_player(player)
-        # if player.hp <= 0:
-        #     print("You have suffered mortal wounds and died. Game over.")
-        #     sys.exit()
-        action_input = get_play_command()
-        if action_input.lower() in ["forward", "for", "fore", "f"]:
-            print("Go forward!")
-            player.move_forward()
-        elif action_input.lower() in ["aftward", "aft", "a", "afterward",
-                                      "aftword"]:
-            print("Go aftward!")
-            player.move_aftward()
-        elif action_input.lower() in ["port", "p", "part", "portward"]:
-            print("Go portside!")
-            player.move_port()
-        elif action_input.lower() in ["starboard", "star board", "star", "s"]:
-            print("Go starboard side!")
-            player.move_starboard()
-        elif action_input.lower() in ["i", "inventory"]:
-            player.print_inventory()
-        elif action_input.lower() in ["a", "attack"]:
-            player.attack()
-        elif action_input.lower() in ["h", "heal"]:
-            player.heal()
-        elif action_input.lower() in ["p", "protect"]:
-            player.protect()
+        choose_action(position, player)
+
+
+def get_available_actions(position, player):
+    """Only make valid actions available. Actions are stored in a dictionary"""
+    actions = OrderedDict()
+    print("Choose an action: ")
+    if player.inventory:
+        action_adder(actions, "i", player.print_inventory, "Print Inventory")
+    if isinstance(position, ship.EnemyTile) and position.enemy.is_alive():
+        action_adder(actions, "a", player.attack, "Attack")
+        action_adder(actions, "p", player.protect, "Protection")
+    else:
+        if ship.tile_at(position.x, position.y - 1):
+            action_adder(actions, "f", player.move_forward, "Go forward")
+        if ship.tile_at(position.x, position.y + 1):
+            action_adder(actions, "a", player.move_aftward, "Go aftward")
+        if ship.tile_at(position.x - 1, position.y):
+            action_adder(actions, "p", player.move_port, "Go portside")
+        if ship.tile_at(position.x + 1, position.y):
+            action_adder(actions, "s", player.move_starboard,
+                         "Go starboard side")
+    if player.hp < 100:
+        action_adder(actions, "h", player.heal, "Heal")
+
+    return actions
+
+
+def action_adder(action_dict, hotkey, action, name):
+    """Add actions to the dictionary"""
+    hotkey = hotkey.lower()
+    action_dict[hotkey] = action
+    print("{}: {}".format(hotkey, name))
+
+
+def choose_action(position, player):
+    """Ask the user to choose an action"""
+    action = None
+    while not action:
+        available_actions = get_available_actions(position, player)
+        action_input = input("Action: ")
+        action = available_actions.get(action_input)
+        if action:
+            action()
         else:
             print("Invalid action!")
-
-
-def get_play_command():
-    """Input command for user """
-    return input("Action: ")
 
 
 play()